refactor: Replace #if-excluded Image and DispatcherAnimator bases
Both base files sit entirely inside a preprocessor branch that is false
for Uno.UI: Image.cs is guarded by !IS_UNIT_TESTS && !UNO_REFERENCE_API
and DispatcherAnimator.cs by !__SKIA__, while Uno.UI.csproj defines
UNO_REFERENCE_API and __SKIA__ and never defines IS_UNIT_TESTS
(Uno.UI.UnitTests consumes Uno.UI through a ProjectReference, so its
constant does not reach these sources). Neither file contributed a line
to any build.

Drop the dead bases and promote the Skia implementations in their place,
which is what those types actually compile to today.

Isolated in its own commit: unlike the rest of the branch this deletes
source rather than moving it, so it can be reverted on its own.
